// src/payload-generated-schema.ts
export const guest_sessions = sqliteTable('guest_sessions', {
  id: integer('id').primaryKey(),
  guestId: text('guest_id').notNull(),
  createdAt: text('created_at').notNull().default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
  lastActiveAt: text('last_active_at').notNull().default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
  expiresAt: text('expires_at').notNull().default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
  locale: text('locale', {enum: ['es', 'en']}),
  theme: text('theme', {enum: ['light', 'dark', 'system']}).default("system"),
  notificationsEnabled: integer('notifications_enabled', {mode: 'boolean'}).default(true),
  integrations: text('integrations', {mode: 'json'}),
  focusSettings: text('focus_settings', {mode: 'json'}),
  updatedAt: text('updated_at').notNull().default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
}, (columns) => [
    uniqueIndex('guest_sessions_guest_id_idx').on(columns.guestId), index('guest_sessions_updated_at_idx').on(columns.updatedAt),
]
)


export const lists = sqliteTable('lists', {
  id: integer('id').primaryKey(),
  name: text('name').notNull(),
  icon: text('icon').default("list"),
  color: text('color'),
  guestId: text('guest_id').notNull(),
  isDefault: integer('is_default', {mode: 'boolean'}).default(false),
  sortOrder: numeric('sort_order', {mode: 'number'}),
  updatedAt: text('updated_at').notNull().default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
  createdAt: text('created_at').notNull().default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
}, (columns) => [
    index('lists_guest_id_idx').on(columns.guestId), index('lists_updated_at_idx').on(columns.updatedAt), index('lists_created_at_idx').on(columns.createdAt),
]
)
// ...
export const tasks = sqliteTable('tasks', {
  id: integer('id').primaryKey(),
  title: text('title').notNull(),
  description: text('description'),
  status: text('status', {enum: ['pending', 'completed']}).notNull().default("pending"),
  important: integer('important', {mode: 'boolean'}).default(false),
  dueDate: text('due_date').default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
  list: integer('list_id').references(() => lists.id, {
      onDelete: 'set null'
  }),
  guestId: text('guest_id').notNull(),
  sortOrder: numeric('sort_order', {mode: 'number'}),
  completedAt: text('completed_at').default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
  updatedAt: text('updated_at').notNull().default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
  createdAt: text('created_at').notNull().default(sql`(strftime('%Y-%m-%dT%H:%M:%fZ', 'now'))`),
}, (columns) => [
    index('tasks_list_idx').on(columns.list), index('tasks_guest_id_idx').on(columns.guestId), index('tasks_updated_at_idx').on(columns.updatedAt), index('tasks_created_at_idx').on(columns.createdAt),
]
)